requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
cryptography>=41.0.0
dash>=2.14.0
plotly>=5.18.0
pandas>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
SQLAlchemy>=2.0.0
python-dateutil>=2.8.2
tqdm>=4.66.0
scikit-learn>=1.4.0
nltk>=3.8.1
textblob>=0.17.1
numpy>=1.26.0
spacy>=3.7.0
transformers>=4.36.0
sentence-transformers>=2.2.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
flask>=2.0.0
flask-cors>=4.0.0